requirements.txt
firebase-admin~=6.8.0
Pyrebase~=3.0.27
python-dotenv~=1.1.0
protobuf~=5.29.4
customtkinter~=5.2.2
reportlab~=4.4.1
requests~=2.32.3
platformdirs~=4.3.8